# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Subcults API
  description: |
    Privacy-first platform for mapping underground music communities.
    Combines real-time location awareness, live audio streaming, and trust-based discovery.

    ## Authentication
    JWT-based authentication with two token types:
    - **Access tokens**: 15-minute expiry, includes `did` (Decentralized Identifier) claim
    - **Refresh tokens**: 7-day expiry, cookie-only

    Pass the access token via `Authorization: Bearer <token>` header or the `access_token` cookie.

    ## Privacy
    All public location data uses geohash-based jitter to prevent precise location tracking.
    Precise coordinates are only available when `allow_precise` consent is granted.

    ## Rate Limiting
    Per-endpoint rate limits are enforced. When exceeded, the API returns `429 Too Many Requests`
    with a `Retry-After` header.
  version: 0.1.0
  contact:
    name: Subcults
  license:
    name: MIT
servers:
  - url: /
    description: Current deployment
...
paths:
...
  /trust/{sceneId}:
    parameters:
      - name: sceneId
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getTrustScore
      tags: [Trust]
      summary: Get the trust score for a scene
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Trust score
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TrustScoreResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Payments ────────────────────────────────────────────────────────
...
  /uploads/sign:
    post:
      operationId: signUpload
      tags: [Uploads]
      summary: Get a pre-signed upload URL for R2
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUploadRequest'
      responses:
        '200':
          description: Pre-signed upload URL
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUploadResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ── LiveKit ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT access token. Include as Authorization: Bearer <token> header or access_token cookie.'
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - validation_error
                - auth_failed
                - not_found
                - rate_limited
                - internal_error
                - forbidden
                - conflict
                - bad_request
                - invalid_palette
                - scene_deleted
                - scene_not_found
                - invalid_scene_name
                - duplicate_scene_name
                - invalid_time_range
                - missing_target
                - unsupported_type
                - invalid_weight
                - alliance_deleted
                - self_alliance
                - already_onboarded
                - not_onboarded
            message:
              type: string
              description: Human-readable error message

    # ── Geographic ──────────────────────────────────────────────────
...
    TrustScoreBreakdown:
      type: object
      properties:
        average_alliance_weight:
          type: number
          format: double
        average_membership_trust_weight:
          type: number
          format: double
        role_multiplier_aggregate:
          type: number
          format: double
    TrustScoreResponse:
      type: object
      properties:
        scene_id:
          type: string
        trust_score:
          type: number
          format: double
        breakdown:
          $ref: '#/components/schemas/TrustScoreBreakdown'
        stale:
          type: boolean
        last_updated:
          type: string

    # ── Membership ──────────────────────────────────────────────────
...
    SignUploadRequest:
      type: object
      required: [contentType, sizeBytes]
      properties:
        contentType:
          type: string
          description: MIME type of the file
        sizeBytes:
          type: integer
          format: int64
          minimum: 1
        postId:
          type: string
    SignUploadResponse:
      type: object
      properties:
        url:
          type: string
          description: Pre-signed S3/R2 upload URL
        key:
          type: string
          description: R2 object key
        expiresAt:
          type: string
          description: Expiration time (ISO 8601)

    # ── LiveKit ─────────────────────────────────────────────────────
...